requirements.txt
dash==0.28.5
dash-html-components==0.13.2
dash-core-components==0.36.0
pandas
attrs
plotly